pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ai-issue-agent"
version = "0.1.0"
description = "Automation system that monitors chat platforms for Python tracebacks and automatically triages them as GitHub issues"
authors = [
    {name = "AI Issue Agent Contributors"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
keywords = ["automation", "github", "slack", "issue-tracking", "error-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Bug Tracking",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "asyncio-atexit>=1.0.1",
    "slack-bolt>=1.18.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
    "openai>=1.6.0",
    "anthropic>=0.8.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "respx>=0.20.0",
    "pytest-timeout>=2.2.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "types-cachetools>=5.3.0",
    "ruff>=0.1.9",
    "pip-audit>=2.6.0",
    "pre-commit>=3.6.0",
    "ipython>=8.19.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
    "pymdown-extensions>=10.7.0",
]

[project.urls]
Homepage = "https://github.com/jtdub/ai-issue-agent"
Repository = "https://github.com/jtdub/ai-issue-agent"
Documentation = "https://github.com/jtdub/ai-issue-agent/blob/main/docs"

[project.scripts]
ai-issue-agent = "ai_issue_agent.__main__:main"

[tool.poetry]
packages = [
    { include = "ai_issue_agent", from = "src" }
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
files = ["src"]

[[tool.mypy.overrides]]
module = [
    "slack_bolt.*",
    "slack_sdk.*",
    "anthropic.*",
    "openai.*",
    "tenacity.*",
    "cachetools.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "S",      # flake8-bandit (security)
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "S101",   # assert used - needed for tests
    "S603",   # subprocess without shell=True - we validate inputs
    "S607",   # partial path to executable - we use list args
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]

[tool.ruff.lint.isort]
known-first-party = ["ai_issue_agent"]

[tool.coverage.run]
source = ["src/ai_issue_agent"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@overload",
    "\\.\\.\\.",
]
fail_under = 80
show_missing = true
